renamed and moved mongoBOConnector according to new pattern
